Finish basic, working version
I basically re-wrote the entire thing because I did not
understand Promises properly. Now I just call .then() a lot
of times and made a lot of tiny functions that do the actual
work, but without any of the Promise nonsense.

As for the config, basically I modify the feed object
throughout the code until, at the end of the script, it
becomes the new config object. This one is then stored as
the new config file.